short form for urls

+++ main/bin/patch.py
@@ -52,14 +52,17 @@ def resave_posts(patt, skip=0, limit=1000):
         print "resaving %s, %s" % (post.id, post.title)
         post.save()
     
-def reduce_notes(weeks=30):
+def reduce_notes(weeks=30, limit=500):
  
     since = datetime.now() - timedelta(weeks=weeks)
-    query = models.Note.objects.filter(date__lt=since)
-    
-    print "*** deleting %s entries" % query.count()
-    
-    query.delete()
+    for user in models.User.objects.all():
+        
+        size = models.Note.objects.filter(target=user).count()
+        if size > limit:
+            query = models.Note.objects.filter(target=user, date__lt=since)
+            dsize = query.count()    
+            print "*** deleting %s entries" % dsize
+            query.delete()
     
     
 @transaction.commit_manually()
